DAY-6 green

+++ src/main/scala/day06/model/RaceData.scala
@@ -1,5 +1,5 @@
 package day06.model
 
-case class RaceData(length: Long, recordDistance: Long) {
+case class RaceData(length: Int, recordDistance: Int) {
 
 }
